+++ MekHQ/unittests/mekhq/campaign/personnel/procreation/AbstractProcreationTest.java
@@ -407,7 +407,7 @@ public class AbstractProcreationTest {
 
         final Person mockPerson = mock(Person.class);
         try (MockedStatic<Compute> compute = Mockito.mockStatic(Compute.class)) {
-            compute.when(Compute::randomInt).thenReturn(0.24f);
+            compute.when(Compute::randomFloat).thenReturn(0.24f);
 
             when(mockPerson.isPregnant()).thenReturn(false);
             mockProcreation.processPregnancyComplications(mockCampaign, LocalDate.ofYearDay(3025, 1), mockPerson);
